+++ src/sparnatural/generators/helpers/ClassBuilder.ts
