+++ src/sparnatural/generators/helpers/IntersectionBuilder.ts
