+++ src/sparnatural/generators/helpers/WhereBuilder.ts
@@ -1,12 +1,12 @@
 import * as DataFactory from "@rdfjs/data-model" ;
 import { BgpPattern, Pattern, Variable } from "sparqljs";
-import { OptionTypes } from "../components/builder-section/groupwrapper/criteriagroup/optionsgroup/OptionsGroup";
-import GroupWrapper from "../components/builder-section/groupwrapper/GroupWrapper";
-import { AbstractWidget } from "../components/widgets/AbstractWidget";
-import ISpecProvider from "../spec-providers/ISpecProviders";
-import ClassBuilder from "./builders/ClassBuilder";
-import IntersectionBuilder from "./builders/IntersectionBuilder";
-import SparqlFactory from "./SparqlFactory";
+import { OptionTypes } from "../../components/builder-section/groupwrapper/criteriagroup/optionsgroup/OptionsGroup";
+import GroupWrapper from "../../components/builder-section/groupwrapper/GroupWrapper";
+import { AbstractWidget } from "../../components/widgets/AbstractWidget";
+import ISpecProvider from "../../spec-providers/ISpecProviders";
+import ClassBuilder from "./ClassBuilder";
+import IntersectionBuilder from "./IntersectionBuilder";
+import SparqlFactory from "../SparqlFactory";
 
 export default class WhereBuilder{
     #resultPtrns: Pattern[] = []
